Rewrite GKDBus exception handling, part 1.
In some situation, current exception handling could lead to memory leaks.
For example, suppose that an instance of RemoteMethodCall is allocated,
and a GKDBusMessage::appendString call fails with OOM, you ends with an
hosed message unfreed.

Implement GKDBusMessage::abandon and friends, and rename GKDBusOOMWrongBuild
exception to GKDBusMessageWrongBuild to get a better handling in such
situation.

Others commits will follow.