Implement directory monitoring using inotify.
This allows us to reload configuration file on inotify event.

If watched directory is deleted or renamed while desktop service is
running, watch is automatically removed. In this scenario, if desktop
service is saving another configuration file (and recreating the
directory in the process), the watch is restored.

deviceConfigurationFile class is removed. The function is merged into
FileSystem class in a more generic way.

This commit is in fact a squashed commit.